megre find-global-declarations.js into acorn-globals.js
