+++ lib/jsdom/acorn-globals.js
@@ -66,6 +66,9 @@ module.exports.parse = reallyParse;
 function findGlobals(source, options) {
   options = options || {};
   const globals = [];
+  const globalDeclarations = [];
+  const globalVariableDeclarations = [];
+  const globalFunctionDeclarations = [];
   let ast;
   const { inBrowser } = options;
   // istanbul ignore else
@@ -78,6 +81,7 @@ function findGlobals(source, options) {
   if (!(ast && typeof ast === "object" && ast.type === "Program")) {
     throw new TypeError("Source must be either a string of JavaScript or an acorn AST");
   }
+  const globalScope = ast;
   const inModule = ast.sourceType === "module";
   const index = ast.body.findIndex(
     statement => statement.type !== "ExpressionStatement" || !statement.directive
@@ -133,6 +137,32 @@ function findGlobals(source, options) {
     ast.locals = ast.locals || Object.create(null);
     ast.locals[node.local.name] = true;
   }
+  function globallyDeclarePattern(node, varNode, declaratorNode) {
+    switch (node.type) {
+      case "Identifier":
+        varNode.globals[node.name] = true;
+        declaratorNode.globals[node.name] = true;
+        break;
+      case "ObjectPattern":
+        node.properties.forEach(property => {
+          globallyDeclarePattern(property.value || property.argument, varNode, declaratorNode);
+        });
+        break;
+      case "ArrayPattern":
+        node.elements.forEach(element => {
+          if (element) {
+            globallyDeclarePattern(element, varNode, declaratorNode);
+          }
+        });
+        break;
+      case "RestElement":
+        globallyDeclarePattern(node.argument, varNode, declaratorNode);
+        break;
+      case "AssignmentPattern":
+        globallyDeclarePattern(node.left, varNode, declaratorNode);
+        break;
+    }
+  }
   const isVarDeclarationScope = !inBrowser || inModule ? isScope : isFunctionScope;
   const isFunctionDeclarationScopeNotInModule = inStrictMode ? isBlockLevelScope : isFunctionScope;
   const isFunctionDeclarationScope = !inBrowser || inModule ? isScope : isFunctionDeclarationScopeNotInModule;
@@ -149,6 +179,16 @@ function findGlobals(source, options) {
         node.declarations.forEach(declaration => {
           declarePattern(declaration.id, parent);
         });
+      } else {
+        node.globals = node.globals || Object.create(null);
+        node.declarations.forEach(declarator => {
+          declarator.globals = declarator.globals || Object.create(null);
+          globallyDeclarePattern(declarator.id, node, declarator);
+        });
+        globalDeclarations.push(node);
+        globalVariableDeclarations.push(node);
+        globalScope.globals = globalScope.globals || Object.create(null);
+        Object.assign(globalScope.globals, node.globals);
       }
     },
     FunctionDeclaration(node, parents) {
@@ -163,6 +203,30 @@ function findGlobals(source, options) {
         if (node.id) {
           parent.locals[node.id.name] = true;
         }
+      } else {
+        node.parent = parents[parents.length - 2];
+        // const index = parents.slice(0, -1).findLastIndex(statement => statement.type !== "LabeledStatement");
+        let i = parents.length - 2;
+        while (i >= 0) {
+          if (parents[i].type !== "LabeledStatement") {
+            break;
+          }
+          --i;
+        }
+        node.scope = i !== -1 ? parents[i] : null;
+        node.outermostLabeledStatement = i < parents.length - 2 ? parents[i + 1] : null;
+        if (node.outermostLabeledStatement) {
+          node.outermostLabeledStatement.id = node.id;
+          node.outermostLabeledStatement.parent = node.scope;
+        }
+        node.globals = node.globals || Object.create(null);
+        if (node.id.name) {
+          node.globals[node.id.name] = true;
+        }
+        globalDeclarations.push(node);
+        globalFunctionDeclarations.push(node);
+        globalScope.globals = globalScope.globals || Object.create(null);
+        Object.assign(globalScope.globals, node.globals);
       }
       declareFunction(node);
     },
@@ -255,7 +319,12 @@ function findGlobals(source, options) {
     groupedGlobals[name] = groupedGlobals[name] || [];
     groupedGlobals[name].push(node);
   });
-  return Object.keys(groupedGlobals).sort().map(name => {
+  const result = Object.keys(groupedGlobals).sort().map(name => {
     return { name, nodes: groupedGlobals[name] };
   });
+  result.globalDeclarations = globalDeclarations;
+  result.globalVariableDeclarations = globalVariableDeclarations;
+  result.globalFunctionDeclarations = globalFunctionDeclarations;
+  result.globalScope = globalScope;
+  return result;
 }
