+++ lib/jsdom/vm-shim.js
@@ -4,7 +4,6 @@ const acorn = require("acorn");
 const findGlobals = require("./acorn-globals.js");
 const escodegen = require("escodegen");
 const jsGlobals = require("./browser/js-globals.json");
-const { findGlobalDeclarations } = require("./find-global-declarations.js");
 
 // We can't use the default browserify vm shim because it doesn't work in a web worker.
 
@@ -242,7 +241,7 @@ exports.runInContext = function (code, contextifiedSandbox, options) {
   escodegen.attachComments(ast, comments, tokens);
 
   const globals = findGlobals(ast, { inBrowser: true });
-  const { globalVariableDeclarations, globalFunctionDeclarations, globalScope } = findGlobalDeclarations(ast);
+  const { globalVariableDeclarations, globalFunctionDeclarations, globalScope } = globals;
 
   // Global "var" declarations and global function declarations defined variables as properties of global object
   // at start of program.
